refactor(core): class-based sensors with engine registerSensor
- Add SensorBase and PointerSensor, KeyboardSensor, KeyboardMoveSensor classes
- Neodrag.registerSensor / unregisterSensor / getSensors with per-key cleanup
- Keep pointerSensor() factory aliases; defaultSensors() returns class instances
