+++ packages/core/src/sensors/index.ts
@@ -1,5 +1,19 @@
-export { pointerSensor } from './pointer.ts';
-export { keyboardSensor } from './keyboard.ts';
-export { keyboardMoveSensor } from './keyboard-move.ts';
+export { SensorBase } from './sensor-base.ts';
+export {
+	PointerSensor,
+	POINTER_SENSOR_KEY,
+	pointerSensor,
+} from './pointer-sensor.ts';
+export {
+	KeyboardSensor,
+	KEYBOARD_SENSOR_KEY,
+	keyboardSensor,
+} from './keyboard-sensor.ts';
+export {
+	KeyboardMoveSensor,
+	KEYBOARD_MOVE_SENSOR_KEY,
+	keyboardMoveSensor,
+	isDraggableKeyboardTarget,
+} from './keyboard-move-sensor.ts';
 export { defaultSensors } from './defaults.ts';
 export type { Sensor, SensorHost, PointerSensorOptions, KeyboardSensorOptions } from './types.ts';
